+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - add CodeFactor.io
 - add snyk.io
 - refactor convert_1_0000_0000_to_9999_9999_9999_9999
+- refactor CharacterSet
 
 ## `1.6.1` - 18 June 2022